+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Remove the previously deprecated `StdError::Unauthorized`. Contract specific
   errors should be implemented using custom error types now (see
   [migration guide](./MIGRATING.md) 0.10 -> 0.11).
+- Use dependency `thiserror` instead of `snafu` to implement `StdError`. Along
+  with this change, the `backtraces` feature now requires Rust nightly.
+- Rename `StdError::ParseErr::source` to `StdError::ParseErr::source_type` and
+  `StdError::SerializeErr::target` to `StdError::SerializeErr::target_type` to
+  work around speacial treatment of the field name `source` in thiserror.
 
 **cosmwasm-vm**